Board has black and white colors and feeds 2D array into draw callback func.
